chore : 디자인 시스템 theme 분리 및 타입 설정
