+++ src/app/styles/theme.ts
@@ -1,71 +1,5 @@
 import { DefaultTheme } from 'styled-components';
 
-declare module 'styled-components' {
-  export interface DefaultTheme {
-    color: {
-      primaryBlue: string;
-      pointPink: string;
-      white: string;
-      black100: string;
-      black80: string;
-      gray50: string;
-      gray30: string;
-      blue50: string;
-      blue30: string;
-      blue10: string;
-    };
-    typo: {
-      title: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-      subtitleLarge: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-      subtitleMedium: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-      bodySemiBold: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-      bodyMedium: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-      captionLarge: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-      captionSmall: {
-        fontFamily: string;
-        fontWeight: number;
-        fontSize: string;
-        lineHeight: string;
-      };
-    };
-    deviceSize: {
-      desktop: string;
-      tablet: string;
-      mobile: string;
-    };
-  }
-}
-
 export const theme: DefaultTheme = {
   color: {
     primaryBlue: '#1D3152',
